+++ src/app/app.component.html
@@ -228,7 +228,7 @@
       <h1>Hello, {{ title }}</h1>
       <p>Congratulations! Your app is running. 🎉</p>
     </div>
-    <div class="divider" role="separator" aria-label="Divider"></div>
+    <hr class="divider" aria-label="Divider">
     <div class="right-side">
       <div class="pill-group">
         @for (item of [
