+++ src/app/app.routes.ts
@@ -1,0 +1,7 @@
+import { Routes } from '@angular/router';
+import { AppComponent } from './app.component';
+
+export const routes: Routes = [
+  { path: '', component: AppComponent },
+  { path: '**', redirectTo: '' }
+];
