--- src/app/home/artifact-preview/artifact-preview.component.html
@@ -1,1 +1,0 @@
-<p>artifact-preview works!</p>
