--- src/app/home/artifact-preview/artifact-preview.component.ts
@@ -1,11 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-    selector: 'app-artifact-preview',
-    imports: [],
-    templateUrl: './artifact-preview.component.html',
-    styleUrl: './artifact-preview.component.css'
-})
-export class ArtifactPreviewComponent {
-
-}
