--- src/app/home/workflow-preview/workflow-preview.component.html
@@ -1,1 +1,0 @@
-<p>workflow-preview works!</p>
