--- src/app/home/workflow-preview/workflow-preview.component.ts
@@ -1,11 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-    selector: 'app-workflow-preview',
-    imports: [],
-    templateUrl: './workflow-preview.component.html',
-    styleUrl: './workflow-preview.component.css'
-})
-export class WorkflowPreviewComponent {
-
-}
