+++ src/main.server.ts
@@ -1,6 +1,35 @@
 import { bootstrapApplication } from '@angular/platform-browser';
+import { provideServerRendering } from '@angular/platform-server';
 import { AppComponent } from './app/app.component';
+import { provideRouter } from '@angular/router';
+import { routes } from './app/app.routes';
+import { provideHttpClient, withInterceptors } from '@angular/common/http';
+import { importProvidersFrom } from '@angular/core';
+import { NgbModule } from '@ng-bootstrap/ng-bootstrap';
+import { TranslateModule, TranslateLoader } from '@ngx-translate/core';
+import { HttpClient } from '@angular/common/http';
+import { TranslateHttpLoader } from '@ngx-translate/http-loader';
 
-const bootstrap = () => bootstrapApplication(AppComponent, config);
+// Factory function for ngx-translate Http loader
+export function HttpLoaderFactory(http: HttpClient) {
+  return new TranslateHttpLoader(http, '/assets/locale/', '.json');
+}
 
-export default bootstrap;
+bootstrapApplication(AppComponent, {
+  providers: [
+    provideServerRendering(), // Habilita SSR
+    provideRouter(routes),
+    provideHttpClient(),
+    importProvidersFrom(
+      NgbModule,
+      TranslateModule.forRoot({
+        loader: {
+          provide: TranslateLoader,
+          useFactory: HttpLoaderFactory,
+          deps: [HttpClient],
+        },
+        defaultLanguage: 'en',
+      })
+    ),
+  ],
+}).catch(err => console.error(err));
